requirements_dev.txt
asyncpg==0.20.1
factory-boy==3.0.1
nyoibo==0.1.0
pre-commit==2.6.0
pytest==5.4.3
pytest-asyncio==0.14.0
pytest-cov==2.10.0
pytest-dotenv-0.5.2
sphinx==3.1.2
sphinx-rtd-theme==0.5.0
uvloop==0.14.0
xenon==0.7.0
